[front] Add users container template and refactor profile
