+++ front/src/containers/header/index.js
@@ -3,6 +3,7 @@ import { Link } from "react-router-dom";
 
 import './style.css';
 
+//Users button should be removed
 export const HeaderContainer = () => {
 
   return (
@@ -10,6 +11,7 @@ export const HeaderContainer = () => {
       <header>
         <Link to="/" className={"btn"}>Articles</Link>
         <Link to="articles/new" className={"btn"}>New Article</Link>
+        <Link to="users" className={"btn"}>Users</Link>
         <Link to="profile" className={"btn"}>Profile</Link>
       </header>
     </>
