+++ front/src/containers/profile/index.js
@@ -1,0 +1,13 @@
+import { Profile } from "../../components/profile";
+
+
+export const ProfileContainer = () => {
+  const userName = "Yurii L"
+  const avatar = "/files/1.jpg";
+
+  return (
+    <>
+      <Profile userName={userName} avatar={avatar}/>
+    </>
+  );
+};
