+++ front/src/containers/users/index.js
@@ -1,0 +1,12 @@
+import { Users } from "../../components/users";
+
+
+export const UsersContainer = () => {
+  const users = [];
+
+  return (
+    <>
+      <Users users={users}/>
+    </>
+  );
+};
