+++ front/src/propsValidation/ProfileValidation_v2.js
@@ -1,0 +1,9 @@
+const PropTypes = require("prop-types");
+const UserValidation = require("./UserValidation");
+
+module.exports = {
+  user: PropTypes.shape(Object.assign(UserValidation,
+    {
+      friends: PropTypes.arrayOf(PropTypes.shape(UserValidation)),
+    })).isRequired
+}
